# api/openapi/flowctl.yaml
openapi: "3.0.3"
info:
  title: FlowCtl API
  description: Production-grade multi-tenant workflow platform
  version: "1.0.0"
  contact:
    name: FlowCtl Team
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/secrets:
    get:
      tags: [secrets]
      summary: List secrets (metadata only)
      operationId: listSecrets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Secret list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    name:
                      type: string
                    scope:
                      type: string
                    created_at:
                      type: string
                      format: date-time
    post:
      tags: [secrets]
      summary: Create or update a secret
      operationId: createSecret
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, value]
              properties:
                name:
                  type: string
                value:
                  type: string
                scope:
                  type: string
                  enum: [tenant, workflow]
                  default: tenant
      responses:
        "201":
          description: Secret created
  /api/v1/cron:
    get:
      tags: [cron]
      summary: List cron schedules
      operationId: listCronSchedules
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cron schedule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CronSchedule"
    post:
      tags: [cron]
      summary: Create cron schedule
      operationId: createCronSchedule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workflow_id, expression]
              properties:
                workflow_id:
                  type: string
                  format: uuid
                expression:
                  type: string
                  description: Cron expression (e.g., "0 9 * * *")
                inputs:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Cron schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CronSchedule"
  /api/v1/audit:
    get:
      tags: [audit]
      summary: List audit logs
      operationId: listAuditLogs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: action
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit log list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditLog"
                  total:
                    type: integer
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    CronSchedule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workflow_id:
          type: string
          format: uuid
        expression:
          type: string
        inputs:
          type: object
        enabled:
          type: boolean
        next_run_at:
          type: string
          format: date-time
        last_run_at:
          type: string
          format: date-time
          nullable: true
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
          format: uuid
          nullable: true
        action:
          type: string
        resource:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time